% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{dc_colors}
\alias{dc_colors}
\title{Function to extract datacamp colors as hex codes}
\usage{
dc_colors(...)
}
\arguments{
\item{...}{Character names of datacamp colors}
}
\description{
Function to extract datacamp colors as hex codes
}
