% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_datacamp.R
\name{dc_pal}
\alias{dc_pal}
\title{DataCamp Color Palettes}
\usage{
dc_pal(name = c("accents", "greys", "accents_light"), named = FALSE)
}
\arguments{
\item{name}{Which of three palettes to choose}

\item{named}{Whether the vector should be named}
}
\description{
DataCamp Color Palettes
}
