% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{dc_palette}
\alias{dc_palette}
\title{Return function to interpolate a datacamp color palette}
\usage{
dc_palette(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in dc_palettes()}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a datacamp color palette
}
