% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{dc_palettes}
\alias{dc_palettes}
\title{Datacamp color palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
dc_palettes
}
\description{
Datacamp color palettes
}
\keyword{datasets}
