% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{dc_reactable_light}
\alias{dc_reactable_light}
\alias{dc_reactable_dark}
\title{Create DataCamp-style reactable interactive table}
\usage{
dc_reactable_light(
  data,
  ...,
  resizable = TRUE,
  filterable = TRUE,
  highlight = TRUE,
  borderless = TRUE,
  striped = TRUE,
  compact = TRUE
)

dc_reactable_dark(
  data,
  ...,
  resizable = TRUE,
  filterable = TRUE,
  highlight = TRUE,
  borderless = TRUE,
  striped = TRUE,
  compact = TRUE
)
}
\arguments{
\item{data}{The data frame to make a table from.}

\item{...}{Arguments passed to \code{\link[reactable]{reactable}}.}

\item{filterable}{Should rows be filterable?}

\item{highlight}{Should hovering over a row highlight it?}

\item{borderless}{Should internal borders be removed?}

\item{striped}{Should alternate rows have a different background color?}

\item{compact}{Should cell padding be reduced to increase data density?}
}
\value{
A \code{reactable} HTML widget.
}
\description{
Creates a reactable interactive table with DataCamp style defaults.
}
\examples{
dc_reactable_light(CO2)
dc_reactable_dark(CO2)
}
\seealso{
\code{\link{reactable_theme_light}}
}
