% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{ioslides_presentation}
\alias{ioslides_presentation}
\title{Datacamp styled ioslides presentation}
\usage{
ioslides_presentation(...)
}
\arguments{
\item{...}{parameters to pass to \link[rmarkdown]{ioslides_presentation}}
}
\description{
Datacamp styled ioslides presentation
}
