% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{moon_reader}
\alias{moon_reader}
\title{Datacamp styled xaringan presentation}
\usage{
moon_reader(...)
}
\arguments{
\item{...}{parameters to pass to \link[xaringan]{moon_reader}}
}
\description{
Datacamp styled xaringan presentation
}
