% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{reactable_theme_light}
\alias{reactable_theme_light}
\alias{reactable_theme_dark}
\title{DataCamp themes for reactable interactive tables}
\usage{
reactable_theme_light(
  color = "#222222",
  background_color = dc_colors_new$beige300,
  base_family = "Poppins",
  row_background_color = dc_colors_new$beige100,
  striped_row_background_color = dc_colors_new$beige200,
  row_highlight_font_weight = "bold",
  filter_input_background_color = dc_colors_new$beige100,
  page_button_hover_background_color = dc_colors_new$beige400
)

reactable_theme_dark(
  color = "#F8F8F2",
  background_color = dc_colors_new$navyDark,
  base_family = "Poppins",
  row_background_color = dc_colors_new$navy,
  striped_row_background_color = dc_colors_new$navyLight,
  row_highlight_font_weight = "bold",
  filter_input_background_color = dc_colors_new$navyDark,
  page_button_hover_background_color = dc_colors_new$blueDark
)
}
\arguments{
\item{color}{Text color.}

\item{background_color}{Background color for the header and footer regions.}

\item{base_family}{Font family for all text.}

\item{row_background_color}{Background color for all rows
(\code{striped = FALSE}) or even numbered rows (\code{striped = TRUE}).}

\item{striped_row_background_color}{Background color for odd numbered rows
when \code{striped = TRUE}.}

\item{row_highlight_font_weight}{Font weight for the highlighted row.}

\item{filter_input_background_color}{Background color the the filter input
textboxes.}

\item{page_button_hover_background_color}{Background color when you hover
over the page number/previous/next buttons.}
}
\value{
An object of class \code{reactableTheme}.
}
\description{
Create a theme for reactable interactive tables that is consistent with
DataCamp style.
}
\examples{
reactable_theme_light()
reactable_theme_dark()
}
