% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_datacamp}
\alias{scale_fill_datacamp}
\title{Fill scale constructor for drsimonj colors}
\usage{
scale_fill_datacamp(palette = "accents", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for drsimonj colors
}
