% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_datacamp.R
\name{theme_datacamp}
\alias{theme_datacamp}
\title{DataCamp Theme}
\usage{
theme_datacamp(...)
}
\arguments{
\item{...}{Additional parameters to pass to \code{theme_gray}}
}
\description{
DataCamp Theme
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = wt, color = as.factor(vs))) +
  geom_point() +
  facet_wrap(~ gear) +
  scale_color_manual(name = 'VS', values = dc_pal()) +
  labs(
    title = 'Mileage vs. Weight by Gears',
    subtitle = "Automobile Data (mtcars)",
    x = "Miles per Gallon",
    y = "Weight"
  ) +
  theme_datacamp()
}
