% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R, R/theme_datacamp.R
\name{theme_datacamp_light}
\alias{theme_datacamp_light}
\alias{theme_datacamp_dark}
\alias{theme_datacamp}
\title{DataCamp ggplot2 theme}
\usage{
theme_datacamp_light(
  base_size = 11,
  base_family = "Poppins",
  subtitle_family = "Poppins Light",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  fill_plot = "#f7f3eb",
  fill_panel = fill_plot,
  fill_strip = dc_colors_new$yellow,
  color_geom = "#444444",
  color_text = "#222222",
  fill_text = color_text,
  color_grid = "#e5e1da",
  ...
)

theme_datacamp_dark(
  base_size = 11,
  base_family = "Poppins",
  subtitle_family = "Poppins Light",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  fill_plot = "#05192d",
  color_text = "#F8F8F2",
  fill_panel = "#05192d",
  color_panel = "transparent",
  fill_strip = colorspace::darken(fill_plot, 0.6),
  fill_legend = colorspace::darken(fill_plot, 0.4),
  color_strip = color_text,
  color_axis = color_text,
  color_grid = colorspace::lighten(fill_plot, 0.5),
  color_geom = color_text,
  fill_geom = color_text,
  fill_text = fill_plot,
  ...
)

theme_datacamp(...)
}
\arguments{
\item{...}{Additional parameters to pass to \code{theme_gray}}
}
\description{
DataCamp ggplot2 theme

DataCamp Theme
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = wt, color = as.factor(vs))) +
  geom_point() +
  facet_wrap(~ gear) +
  scale_color_manual(name = 'VS', values = dc_pal()) +
  labs(
    title = 'Mileage vs. Weight by Gears',
    subtitle = "Automobile Data (mtcars)",
    x = "Miles per Gallon",
    y = "Weight"
  ) +
  theme_datacamp()
}
